Moved TimerBars to own Folder and edited a bit.
